+++ app/src/main/java/com/forrestguice/suntimeswidget/themes/defaults/DarkTheme.java
@@ -24,7 +24,6 @@ import android.support.v4.content.ContextCompat;
 import com.forrestguice.suntimeswidget.BuildConfig;
 import com.forrestguice.suntimeswidget.R;
 import com.forrestguice.suntimeswidget.themes.SuntimesTheme;
-import com.forrestguice.suntimeswidget.themes.ThemeBackground;
 
 public class DarkTheme extends SuntimesTheme
 {
