+++ app/src/main/java/com/forrestguice/suntimeswidget/themes/defaults/DarkThemeTranslucent.java
@@ -23,7 +23,6 @@ import android.support.v4.content.ContextCompat;
 
 import com.forrestguice.suntimeswidget.BuildConfig;
 import com.forrestguice.suntimeswidget.R;
-import com.forrestguice.suntimeswidget.themes.ThemeBackground;
 
 public class DarkThemeTranslucent extends DarkThemeTrans
 {
